rename location package to spatial
