+++ info.limpet.rcp/src/info/limpet/rcp/editors/OperationsLibrary.java
@@ -18,8 +18,8 @@ import info.limpet.data.operations.MultiplyQuantityOperation;
 import info.limpet.data.operations.SimpleMovingAverageOperation;
 import info.limpet.data.operations.SubtractQuantityOperation;
 import info.limpet.data.operations.UnitConversionOperation;
-import info.limpet.data.operations.location.BearingBetweenTracksOperation;
-import info.limpet.data.operations.location.DistanceBetweenTracksOperation;
+import info.limpet.data.operations.spatial.BearingBetweenTracksOperation;
+import info.limpet.data.operations.spatial.DistanceBetweenTracksOperation;
 import info.limpet.rcp.analysis_view.AnalysisView;
 import info.limpet.rcp.data_frequency.DataFrequencyView;
 import info.limpet.rcp.operations.ShowInNamedView;
