+++ info.limpet/META-INF/MANIFEST.MF
@@ -35,8 +35,7 @@ Export-Package: com.thoughtworks.xstream,
  info.limpet.data.impl.samples,
  info.limpet.data.math,
  info.limpet.data.operations,
- info.limpet.data.operations.location,
- info.limpet.data.store,
+ info.limpet.data.operations.spatial,
  info.limpet.data.persistence.xml,
  info.limpet.data.store,
  org.xmlpull.mxp1,
